changed dp type to real64 in parametersmod.f90

+++ src/parametersmod.f90
@@ -15,7 +15,7 @@ integer, parameter :: i2 = int16   ! 2 byte integer
 integer, parameter :: i4 = int32   ! 4 byte integer
 integer, parameter :: i8 = int64   ! 8 byte integer
 integer, parameter :: sp = real32  ! 4 byte real
-integer, parameter :: dp = real32  ! 8 byte real
+integer, parameter :: dp = real64  ! 8 byte real
 
 !size
